Create basic website format and personal information
With this first edit, the basic format of the website is established: there is a vertical navigation bar containing my photo; name; links to my project and resume pages; and links to my various social media profiles. Moreover, much of my personal and educational information was added on the 'About Me' page, and my resume PDF was embedded onto the resume page. TODO: Add projects to the projects page.