Fix visuals for wider displays
Several issues were addressed. Firstly, the default text size was increased to 20px so the website is more readable. Second, the body background color was made to be the same as that of main container to prevent main container from being visible. Thirdly, the height of the embedded resume was increased. TODO: Add projects, make column-lists (e.g. courses, technologies) closer.